Add files via upload
this is a live geo-visualization dashboard that displays
information and insights from the Phonepe pulse Github repository in an interactive
and visually appealing manner. The dashboard will have at least 10 different
dropdown options for users to select different facts and figures to display. The data
will be stored in a MySQL database for efficient retrieval and the dashboard will be
dynamically updated to reflect the latest data.
Users will be able to access the dashboard from a web browser and easily navigate
the different visualizations and facts and figures displayed. The dashboard will
provide valuable insights and information about the data in the Phonepe pulse
GitHub repository, making it a valuable tool for data analysis and decision-making.